Use more precise types
Replace the Token interface with Plain because we need to use claims(),
which only exist on Plain anyway. Getting any other token now throws
errors.
Also assert that our token type, AccessToken is returned in
getAccessToken.

This helps with static analysis tools and IDE autocomplete both for us
and users.